HPO_containerized/components/train_model/runtime-requirements.txt
# Generated by KFP.
cloudml-hypertune==0.1.0.dev6
git+https://github.com/jindrvo1/blackfriday
pandas>=2.2.3
scikit-learn>=1.5.2
xgboost>=2.1.2